Refactor: DRY up code w/ princess saveable module Remove repeating methods by adding them to a module, princess_saveable, then including that module in the nexessary files.
